adding io and moving types
